Introduced `mini-face` integration
