--- child_lab_framework/task/gaze/mini_face/__init__.py
@@ -1,3 +1,0 @@
-from .mini_face import Estimator, Result
-
-__all__ = ['Estimator', 'Result']
